save_as_first

+++ Core/ALL.h
@@ -30,8 +30,8 @@
 #include "uart_device.h"
 #include "stdio.h"
 extern uint8_t RunFlag;
-extern int16_t LeftPwm,RightPwm;
-extern int16_t AvePwm,DifPwm; 
+extern float LeftPwm,RightPwm;
+extern float AvePwm,DifPwm;
 extern float Angle;
 extern float GyroY_Actual;
 #endif
